🔧 Include confirmation dialog

+++ app/javascript/components/Recipe.jsx
@@ -24,24 +24,26 @@ const Recipe = () => {
   };
 
   const deleteRecipe = () => {
-    const url = `/api/v1/destroy/${params.id}`;
-    const token = document.querySelector('meta[name="csrf-token"]').content;
+    if (window.confirm("Êtes-vous sûr de vouloir supprimer cette recette ?")) {
+      const url = `/api/v1/destroy/${params.id}`;
+      const token = document.querySelector('meta[name="csrf-token"]').content;
 
-    fetch(url, {
-      method: "DELETE",
-      headers: {
-        "X-CSRF-Token": token,
-        "Content-Type": "application/json",
-      },
-    })
-      .then((response) => {
-        if (response.ok) {
-          return response.json();
-        }
-        throw new Error("Network response was not ok.");
+      fetch(url, {
+        method: "DELETE",
+        headers: {
+          "X-CSRF-Token": token,
+          "Content-Type": "application/json",
+        },
       })
-      .then(() => navigate("/recipes"))
-      .catch((error) => console.log(error.message));
+        .then((response) => {
+          if (response.ok) {
+            return response.json();
+          }
+          throw new Error("Network response was not ok.");
+        })
+        .then(() => navigate("/recipes"))
+        .catch((error) => console.log(error.message));
+    }
   };
 
   const ingredientList = () => {
